cognibrew-cloud-api-gateway/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
websockets==14.2
pydantic==2.11.3
pydantic-settings==2.9.1
python-json-logger==3.3.0
pika==1.3.2
protobuf==6.32.0
PyJWT==2.10.1
